Updated statuses and locations based on emailed symphony policy file
Long term this will be managed outside of code changes. In the preparation
for this Elaine shared the symphony policy file which contains the original
for the mappings that we use in this project (and Discovery).  At this time
Weiwei realized that some of these mappings were missing and potentially
contributing to the 500 errors we see in NEOSDiscovery.