fix "The Economist" rollbar error, short_code normalization, and an undeclared e (#375)
I suspect this was the priority NEOSDiscovery performance error.
catalog/1698093 demonstrated two separate but related bugs.

The first bug was that HoldingsHelper#links repeated the call to Symphony
for each of the items in the holdings table rather than using the @urls
instance variable declared in the controller.

The second bug was that HoldingsHelper#links assumes that there always are
links when (for example the Economist) there may not be any.

While investigating this there appeared to be an inconsistency in the
normalization of Location short_codes.  I have migrated the Location short codes
so that there isn't any conversion by the application.  This matches Discovery.

Additionally this removes an undeclared var e in on of the rollbar messages.