change log_level from debug to warn (#412)
Neil says, "In Production, I've had to dedicate more than 5Gb of disk on each server to /var/ for application logging, and it's still not enough: we're running out of space in the middle of the 3rd week.   I've reviewed the logs & found we're getting debug (!) level logging, which seems like a lot for a Prod environment."

Sam changed Discovery from info to warn 5 years ago.

The rails advice is to "use the lowest log level to ensure availability of diagnostic information when problems arise." Historically we've not needed debug level logs for NEOSDiscovery and Rollbar will get us most of what we need so I don't see the harm.