CirculationRules and ItemType models to align with Discovery
Both Discovery and NEOSDiscovery use Symphony configuration data to provide human readable string in the holding table.  Additionaly NEOSDiscovery has library/location relationships.  We now have this data flowing into Discovery but we need this to continue to flow into NEOSDiscovery.  One way to do this would be to use the same migrations on both applications.

Here we create two tables that will not be used by the application in the interest of being compatible with Discovery symphony_nightly migrations.